Guard against nonexistent indexes before trying to remove
Test Plan: N/A

Reviewed-on: https://gerrit.instructure.com/130984
Tested-by: Jenkins
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>

+++ db/migrate/20170815211443_add_assignment_group_id_to_scores.rb
@@ -45,11 +45,15 @@ class AddAssignmentGroupIdToScores < ActiveRecord::Migration[5.0]
               algorithm: :concurrently,
               name: :index_course_scores
 
-    remove_index :scores, column: :enrollment_id,
-                 name: :index_scores_on_enrollment_id
+    if index_exists?(:scores, :enrollment_id, name: "index_scores_on_enrollment_id")
+      remove_index :scores, column: :enrollment_id,
+                   name: :index_scores_on_enrollment_id
+    end
 
-    remove_index :scores, column: [:enrollment_id, :grading_period_id],
-                 name: :index_scores_on_enrollment_id_and_grading_period_id
+    if index_exists?(:scores, [:enrollment_id, :grading_period_id], name: "index_scores_on_enrollment_id_and_grading_period_id")
+      remove_index :scores, column: [:enrollment_id, :grading_period_id],
+                   name: :index_scores_on_enrollment_id_and_grading_period_id
+    end
 
     reversible do |direction|
       direction.down { DataFixup::DeleteScoresForAssignmentGroups.run }
